feat: uncheck sets LastPurch a day ago (so that it stays that way).
